+++ src/graph/Stockage/StockageType.java
@@ -12,7 +12,7 @@ public interface StockageType {
     void init(int V);
     void addEdge(Edge e);
     void addEdge(Vertex v, Vertex w);
-    void addEdge(Vertex v, Vertex w, int weigth);
+    void addEdge(Vertex v, Vertex w, double weigth);
     Edge getEdge(Vertex v, Vertex w);
     LinkedList<Vertex> adjacentVertex(Vertex v);
     LinkedList<Edge> adjacentEdges(Vertex v);
